backend/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
supabase==2.28.0
Pillow==10.4.0